requirements.txt
pypdf
langchain
chromadb # Vector storage
pytest
pytest-asyncio
langchain-openai
langchain-community
langchain-chroma 
python-dotenv
ragas